requirements.txt
aniso8601==9.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
docker==5.0.3
Flask==2.1.2
Flask-JWT==0.3.2
Flask-JWT-Extended==4.4.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
pywin32==227
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.10
websocket-client==1.3.3
Werkzeug==2.1.2
zipp==3.8.0
